Add basic member code completion for TypeScript.
Pressing the dot character will show the code completion window with
a list of members.

Currently only working for the first TypeScript file that the user
types a dot character into. Typing in a dot character into another
TypeScript file does not show a completion window until it is typed
in again. Looks like the ILanguageServiceShimHost needs to have
all the TypeScript files in its script list before
TypeScriptServicesFactory.createLanguageServiceShim() is called.
If this is not done then dot completion only works after the
script version is incremented. Using the language service refresh()
method does not seem to fix this problem.