Implement basic Find References for TypeScript.
Add Find TypeScript References menu item to the Search menu
and the text editor's right click context menu.

Unable to override the standard Find References menu item in
a simple way since it relies on the SharpDevelop DOM which
is not being used by the TypeSript addin.

Currently only supports references in a single file.