Add code folding/quick browse support for classes.
Class members are not currently supported.

Currently using SharpDevelop's parser service and a custom
parser class which generates dummy classes based on the
outlining regions that TypeScript returns. Unfortunately
this causes a side effect of adding any TypeScript classes
to the C# code completion. In the future look to prevent
the unwanted side effect.