Do not use the SharpDevelop parser service for TypeScript.
To prevent TypeScript classes from appearing in C# or
VB.NET code completion the SharpDevelop parser service
is no longer used when generating folding for the
text editor. Instead a separate TypeScriptParserService
that runs in the background and updates the active
text editor's folding and quick class browser if it
contains TypeScript.