Match files in TypeScript context ignoring case.
Somehow the bookmarks in the user profile were saved using
a lower case drive letter which does not match the filenames
returned by a project where the drive letter is upper case.
This was causing the TypeScript context to add the same
file twice.

Unable to reproduce how the bookmarks were saved with
a lower case drive letter.