Initial port to MonoDevelop - Xamarin Studio.
Support code completion and method completion for a TypeScript
file inside a project.

Code completion is broken when editing a new TypeScript file that
is not saved outside of a project.

Not implemented:

Compiling to JavaScript.
TypeScript options panel needs to be ported to GTK#.
Code folding.
Adding new project items after compiling to JavaScript.