Improve speed for first completion list to be displayed.
The first time TypeScript's getCompletionsAtPosition method is called
it can take 1 to 2 seconds to complete. Instead of making the user
wait the first time they trigger completion in the text editor the
addin makes a call to getCompletionsAtPosition when the project is
opened.

A better solution would be to move the completion to a background
thread so the project load is not slowed down.