Use the LanguageServiceCompiler when building the project.
Compile all TypeScript files using the LanguageServiceCompiler before
the project is built. Using the LanguageServiceCompiler with an
existing TypeScriptContext, which has already processed all the files
in the project, makes compilation of a single file on save much faster
than before. Previously a new TypeScript compiler was created with
all the project files added to it, including the lib.d.ts file, and
then the overall compilation would take over a second.