Add support for generation declaration files - not currently working.
New option "Generate declaration files" in project options.

Compiler now tries to generate declarations using emitDeclarations
which is available from the LanguageServicesCompiler. This does not
generate a .d.ts file. Debugging the JavaScript the problem seems to be
that the Ast type cannot find the symbol. So there must be some extra
step that is not being currently done in order to get this to work.

DeclarationEmitter.prototype.canEmitDeclarations = function (declAST) {
   var container = DeclarationEmitter.getEnclosingContainer(declAST);
   if (container.kind() === 130 /* ModuleDeclaration */ || container.kind() === 120 /* SourceUnit */) {
     var pullDecl = this.semanticInfoChain.getDeclForAST(declAST);
     if (!TypeScript.hasFlag(pullDecl.flags, 1 /* Exported */)) {
       var start = new Date().getTime();
       var declSymbol = this.semanticInfoChain.getSymbolForAST(declAST);
       var result = declSymbol && declSymbol.isExternallyVisible();
       TypeScript.declarationEmitIsExternallyVisibleTime += new Date().getTime() - start;

       //declSymbol is null
       return result;
     }
  }
}