Fix exception in quick class browser.
The quick class browser was using the
DefaultTypeDefinition.UnresolvedFile property which was not set.

Unfortunately the quick class browser does not show class members.
So the quick class browser integration is disabled. The quick class
browser seems to be tied to the current parser information for the
project, but the TypeScript parse information is separate from the
class parse information. The TypeScript parse information is separate
since TypeScript types should not be available for the project when
coding in C#. The JavaScript addin had this problem since it was
using the standard ParserService.