DEL Typescript service types in favor of  TLS #5
In oder to use an extern TypeScript language service we delete all related files.
we will then include the project CSharpTypescriptLanugageService or
the related nugget.

Reason:
Why to exclude typescript?
Observed in the past, it is like the the Typescript interface change.
It is hard to identify which variable names, function calls was added or changed.

See 1.3 -> 1.4 -> 1.5.

- CSharpTypescriptLanugageService uses V8dotnet,Typescript and T4
templates to transpile the TypeScript interfaces/classes to C#
interfaces/classes.
- it also generates V8 alike function calls to access the V8 context.

Delete the following file types:

- TypeScript interface types which reflect "compiler/types.ts" and
  services/services.ts
- delete helper classes used for marshaling return values of
  "JSON.stringify()".
- delete javascript helper files like "completion.js".

This commit shows also renamed files as deleted like:
LanguageServiceShimHost to LanguageServiceHost which implements
etc.