Undo impl of static v8engine and add ScriptSnapshot #5 #8
This commit withdraw's changes that were introduced do to an false assumption
about TypeScriptContext.cs.
We removed V8TypescriptProvider.cs and placed the V8Engine back to the
TypeScriptContext.cs to restore the old behavior.

To not to introduce V8Engine into the Typescript-addin we added a new
interace (IScriptSnapshotExtension) which adds properties to
IScriptSnapshot. The properties are needed to compare the changes from
an old and current snapshot.