Use custom rename dialog.
MonoDevelop's built in question dialog was previously used via
MessageService.GetTextResponse(...) but this method causes a critical
GTK error.

 Gtk-Critical: IA__gtk_widget_grab_default: assertion
'gtk_widget_get_can_default (widget)' failed

This critical GTK error does not cause a problem on Mac or Windows but
on Linux the dialog would not be displayed. To workaround this a
custom rename dialog is now used.