Use custom rename dialog.
MonoDevelop's built in question dialog was previously used via
MessageService.GetTextResponse(...) but this method causes a critical
GTK error.

 Gtk-Critical: IA__gtk_widget_grab_default: assertion
'gtk_widget_get_can_default (widget)' failed

To workaround this error a custom rename dialog is now used. This
fixes the critical error but on Linux the rename dialog is shown if
the keyboard shortcut F2 is used but not if the Rename context menu is
used.