Fix code completion not working in for..of
Code completion inside "for..of" would not work until the project was
compiled. The variable defined by the for..of statement would produce
no member completion results when the dot character was pressed inside
the for..of loop.

Now when the project is created the compiler settings are updated in
the shim host.