Ensure language service host updated when project options changed.
Previously the language service host's compiler options were not
updated when the project changed. This could cause code completion
to stop working depending on whether the project's ES version was
changed. One example of this is the following code:

for (var foo of "bar") {
    foo.
}

With ES6 the above is invalid and code completion inside the for
loop for foo would not show anything. Switching to ES5 or ES3 would
not fix the code completion not working since this information was
not updated in the language service host. With this commit the code
completion now works when switching to ES5 or ES3 with the above for
of loop.