Fix diagnostic errors not shown in Errors windows and text editor.
The IDocument was created by the parser on a background thread but
the SDTask needs to be created on the UI thread and the document
is then accessed from the UI thread which is disallowed by the
document. So now the real document open in SharpDevelop is used
since that was created on the UI thread.