Reload generated files on saving when compile on save is enabled.
With compile on save enabled when the TypeScript file was saved any
generated files, such as JavaScript files, were not reloaded
automatically if they were open in the text editor. When building the
project the generated files were reloaded.

Now after the TypeScript file is saved the documents open in the
workbench are checked and reloaded if a document is a generated file.
If the generated file is marked as dirty, so it has been edited in
the text editor and has unsaved changes, then the generated file will
not be reloaded.

Fixes #14