Fix warnings when installing addin.
When installing the addin the install would work but show a dialog
with:

The installation has completed with warnings.

The add-in 'MonoDevelop.TypeScript,0.9' is trying to extend '', but
there isn't any add-in defining this extension point
The add-in 'MonoDevelop.TypeScript,0.9' is trying to extend
'/TypeScript', but there isn't any add-in defining this extension point

This was because the wrong attribute was being used for the
MonoDevelop/ProjectModel/LanguageBindings extension.

Fixes #15